refactor: move startWorkerMode to worker.ts, run lifecycle inside main()
startWorkerMode now lives in worker.ts where it belongs — it sets up the
workspace, WorkerSession, and signal handlers, then returns { session, cleanup }
without calling main(). main() calls startWorkerMode internally if a
workerConfig is provided, owns the full lifecycle (setup → loop → cleanup),
and exits after cleanup in worker mode.

This eliminates the circular dependency that forced all worker setup code into
index.ts: worker.ts has zero imports from index.ts. The if/else startup fork
at the bottom of index.ts is gone — a single await main(...) handles both modes,
with one banner printed once inside main(). SIGINT/SIGTERM handlers now live in
startWorkerMode (worker.ts). WS_PROMPT sentinel is hidden behind
WorkerSession.isWsSignal().
